+++ apps/generate-ui-v2/src/components/fields/autocomplete/intellij-autocomplete-styles.ts
