+++ apps/generate-ui-v2/src/components/fields/autocomplete/vscode-autocomplete-styles.ts
@@ -159,59 +159,3 @@ export const vscodeComboboxStyles = () => css`
     flex: 0 0 auto;
   }
 `;
-
-export const vscodeOptionStyles = () => css`
-  ${display('inline-flex')} :host {
-    font-family: var(--body-font);
-    border-radius: ${cornerRadius};
-    border: calc(${borderWidth} * 1px) solid transparent;
-    box-sizing: border-box;
-    color: var(--vscode-foreground);
-    cursor: pointer;
-    fill: currentcolor;
-    font-size: var(--vscode-font-size);
-    line-height: normal;
-    margin: 0;
-    outline: none;
-    overflow: hidden;
-    padding: 0 calc((${designUnit} / 2) * 1px) calc((${designUnit} / 4) * 1px);
-    user-select: none;
-    white-space: nowrap;
-  }
-  :host(:${focusVisible}) {
-    border-color: var(--vscode-focusBorder);
-    background: var(--vscode-list-activeSelectionBackground);
-    color: var(--vscode-foreground);
-  }
-  :host([aria-selected='true']) {
-    background: var(--vscode-list-activeSelectionBackground);
-    border: calc(${borderWidth} * 1px) solid var(--vscode-focusBorder);
-    color: var(--vscode-list-activeSelectionForeground);
-  }
-  :host(:active) {
-    background: var(--vscode-list-activeSelectionBackground);
-    color: var(--vscode-list-activeSelectionForeground);
-  }
-  :host(:not([aria-selected='true']):hover) {
-    background: var(--vscode-list-activeSelectionBackground);
-    border: calc(${borderWidth} * 1px) solid var(--vscode-focusBorder);
-    color: var(--vscode-list-activeSelectionForeground);
-  }
-  :host(:not([aria-selected='true']):active) {
-    background: var(--vscode-list-activeSelectionBackground);
-    color: var(--vscode-foreground);
-  }
-  :host([disabled]) {
-    cursor: ${disabledCursor};
-    opacity: 0.4;
-  }
-  :host([disabled]:hover) {
-    background-color: inherit;
-  }
-  .content {
-    grid-column-start: 2;
-    justify-self: start;
-    overflow: hidden;
-    text-overflow: ellipsis;
-  }
-`;
